+++ client/src/io/worxbend/gitea4s/http/SingleCommitParams.scala
@@ -1,0 +1,10 @@
+package io.worxbend.gitea4s.http
+
+final case class SingleCommitParams(
+    stat: Option[Boolean] = None,
+    verification: Option[Boolean] = None,
+    files: Option[Boolean] = None
+)
+
+object SingleCommitParams:
+  val default: SingleCommitParams = SingleCommitParams()